requirements.txt
absl-py==0.10.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.2.0
audiomentations==0.12.1
audioread==2.1.8
backcall==0.2.0
bleach==3.2.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
configparser==5.0.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
docker-pycreds==0.4.0
docopt==0.6.2
entrypoints==0.3
future==0.18.2
gitdb==4.0.5
GitPython==3.1.8
google-auth==1.21.3
google-auth-oauthlib==0.4.1
grpcio==1.32.0
hydra-core==1.0.3
idna==2.10
importlib-resources==3.0.0
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jiwer==2.1.0
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.1
kiwisolver==1.2.0
librosa==0.8.0
llvmlite==0.34.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
nbclient==0.5.0
nbconvert==6.0.6
nbformat==5.0.7
nest-asyncio==1.4.0
notebook==6.1.4
num2words==0.5.10
numba==0.51.2
numpy==1.19.2
oauthlib==3.1.0
omegaconf==2.0.2
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
pooch==1.2.0
prometheus-client==0.8.0
promise==2.3
prompt-toolkit==3.0.7
protobuf==3.13.0
psutil==5.7.2
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytorch-lightning==0.9.0
PyYAML==5.3.1
pyzmq==19.0.2
requests==2.24.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.2
Send2Trash==1.5.0
sentry-sdk==0.17.8
shortuuid==1.0.1
six==1.15.0
smmap==3.0.4
SoundFile==0.10.3.post1
subprocess32==3.5.4
tb-nightly==2.3.0a20200518
tensorboard==2.2.0
tensorboard-plugin-wit==1.7.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
torch==1.6.0
torchaudio==0.6.0
torchvision==0.7.0
tornado==6.0.4
tqdm==4.49.0
traitlets==5.0.4
typing-extensions==3.7.4.3
urllib3==1.25.10
wandb==0.10.2
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
youtokentome==1.0.6
